Ready app for deploying on GitHub Pages (#9)
In order to deploy the app on GitHub Pages, a duplicate index.html was created in the root directory, with relative paths to site.css and app.js replaced with absolute paths, starting with e.g. /sevenguis/public/...

the `public/js` folder was in gitignore (I know not why), and this was preventing GitHub Pages from finding it. I removed it from gitignore.