Bug modal on non chrome browsers (#13)
* Modal windows now load on Firefox and Safari

Issues with positioning the modals remain. I'm using the HTML dialog element, which is new and not supported by Safari or Firefox. Therefore I used the "dialog polyfill" library from Chrome to enable Safari and Firefox to load it. I attempted to import it with cljsjs, but it did not work (there is a failing test listed on their github repo; that may be related). Instead I simply put the dialog-polyfill.js file in public/js, and added it to public/index.html. (I had to remember to add it also to the duplicate index.html in the root directory that Github Pages uses.)

In the code itself, using dialog polyfill required adding a call to the dialogPolyfill.registerDialog function for each modal ref, with argument the (dereferenced) modal ref.

There is also a dialog-polyfill.css file added, and here is where some of the remaining bugs with positioning may lie. Removing the "position: absolute" entry in dialog-polyfill.css makes the modal work properly in Chrome again, but breaks it in Safari and Firefox entirely (the modal is non-responsive). But putting "position: absolute" back in makes the positioning wrong on Chrome, for a reason I don't understand. I added some hacky positioning to make the Chrome positioning more or less right, and called it a day.